refactor(event-creation): integrate CreateEventForm directly into CreateNewEventDialog, removing the wrapper component
