+++ apps/epicdev-ai/src/app/(content)/events/_components/create-new-event-dialog.tsx
@@ -1,9 +1,15 @@
 'use client'
 
 import * as React from 'react'
+import { useRouter } from 'next/navigation'
+import { createEvent, createEventSeries } from '@/lib/events-query'
+import { api } from '@/trpc/react'
+import { z } from 'zod'
 
+import { type EventCreationResult } from '@coursebuilder/core'
 import {
 	Button,
+	CreateEventForm,
 	Dialog,
 	DialogContent,
 	DialogDescription,
@@ -14,8 +20,7 @@ import {
 } from '@coursebuilder/ui'
 import type { ButtonProps } from '@coursebuilder/ui/primitives/button'
 import { cn } from '@coursebuilder/ui/utils/cn'
-
-import CreateEventFormWrapper from './create-event-form-wrapper'
+import { getResourcePath } from '@coursebuilder/utils-resource/resource-paths'
 
 export default function CreateNewEventDialog({
 	buttonLabel = 'Create new event',
@@ -27,6 +32,29 @@ export default function CreateNewEventDialog({
 	variant?: ButtonProps['variant']
 }) {
 	const [open, setOpen] = React.useState(false)
+	const router = useRouter()
+	const { data: tags } = api.tags.getTags.useQuery()
+	const parsedTags = z
+		.array(
+			z.object({
+				id: z.string(),
+				fields: z.object({
+					label: z.string(),
+					name: z.string(),
+				}),
+			}),
+		)
+		.parse(tags || [])
+
+	const handleSuccess = async (result: EventCreationResult) => {
+		if (result.type === 'single' && result.event) {
+			router.push(getResourcePath('event', result.event.fields?.slug, 'edit'))
+		} else if (result.type === 'series' && result.eventSeries) {
+			router.push(
+				getResourcePath('event', result.eventSeries.fields?.slug, 'edit'),
+			)
+		}
+	}
 
 	return (
 		<div>
@@ -52,7 +80,17 @@ export default function CreateNewEventDialog({
 							for easier management.
 						</DialogDescription>
 					</DialogHeader>
-					<CreateEventFormWrapper />
+					<CreateEventForm
+						tags={parsedTags}
+						onSuccess={handleSuccess}
+						createEvent={createEvent}
+						createEventSeries={createEventSeries}
+						allowMultipleEvents={true}
+						allowCoupons={true}
+						defaultTimezone="America/Los_Angeles"
+						defaultPrice={250}
+						defaultQuantity={40}
+					/>
 				</DialogContent>
 			</Dialog>
 		</div>
